+++ app/nexus/internal/initialization/recovery/recovery.go
@@ -11,13 +11,13 @@ import (
 	"time"
 
 	"github.com/spiffe/go-spiffe/v2/workloadapi"
+	"github.com/spiffe/spike-sdk-go/config/env"
 	"github.com/spiffe/spike-sdk-go/crypto"
 	"github.com/spiffe/spike-sdk-go/log"
 	"github.com/spiffe/spike-sdk-go/retry"
 	"github.com/spiffe/spike-sdk-go/security/mem"
 	"github.com/spiffe/spike-sdk-go/spiffe"
 
-	"github.com/spiffe/spike/app/nexus/internal/env"
 	state "github.com/spiffe/spike/app/nexus/internal/state/base"
 )
 
@@ -67,7 +67,7 @@ func InitializeBackingStoreFromKeepers(source *workloadapi.X509Source) {
 	)
 	defer cancel()
 
-	_, err := retry.Do(ctx, func() (bool, error) {
+	_, err := retry.Forever(ctx, func() (bool, error) {
 		log.Log().Info(fName, "message", "retry:"+time.Now().String())
 
 		initSuccessful := iterateKeepersAndInitializeState(
@@ -83,12 +83,7 @@ func InitializeBackingStoreFromKeepers(source *workloadapi.X509Source) {
 			"keepersSoFar", len(successfulKeeperShards),
 		)
 		return false, ErrRecoveryRetry
-	},
-		retry.WithBackOffOptions(
-			retry.WithMaxInterval(env.RecoveryOperationMaxInterval()),
-			retry.WithMaxElapsedTime(0), // Retry forever.
-		),
-	)
+	})
 
 	// This should never happen since the above loop retries forever:
 	if err != nil {
@@ -138,14 +133,14 @@ func RestoreBackingStoreFromPilotShards(shards []ShamirShard) {
 
 	log.Log().Info(fName,
 		"message", "Recovering backing store using pilot shards",
-		"threshold", env.ShamirThreshold(),
+		"threshold", env.ShamirThresholdVal(),
 		"len", len(shards),
 	)
 
 	// Ensure we have at least the threshold number of shards
-	if len(shards) < env.ShamirThreshold() {
+	if len(shards) < env.ShamirThresholdVal() {
 		log.Log().Error(fName, "message", "Insufficient shards for recovery",
-			"provided", len(shards), "required", env.ShamirThreshold())
+			"provided", len(shards), "required", env.ShamirThresholdVal())
 		return
 	}
 
@@ -178,7 +173,7 @@ func RestoreBackingStoreFromPilotShards(shards []ShamirShard) {
 
 	// Don't wait for the next cycle in `SendShardsPeriodically`.
 	// Send the shards asap.
-	sendShardsToKeepers(source, env.Keepers())
+	sendShardsToKeepers(source, env.KeepersVal())
 }
 
 // SendShardsPeriodically distributes key shards to configured keeper nodes at
@@ -199,7 +194,7 @@ func SendShardsPeriodically(source *workloadapi.X509Source) {
 
 	log.Log().Info(fName, "message", "Will send shards to keepers")
 
-	ticker := time.NewTicker(env.RecoveryKeeperUpdateInterval())
+	ticker := time.NewTicker(env.RecoveryKeeperUpdateIntervalVal())
 	defer ticker.Stop()
 
 	for range ticker.C {
@@ -211,8 +206,8 @@ func SendShardsPeriodically(source *workloadapi.X509Source) {
 			continue
 		}
 
-		keepers := env.Keepers()
-		if len(keepers) < env.ShamirShares() {
+		keepers := env.KeepersVal()
+		if len(keepers) < env.ShamirSharesVal() {
 			log.FatalLn(fName + ": not enough keepers")
 		}